feat(novel_shelf): improve novel info display with new card
新しいNovelInfoCardウィジェットを導入し、ノベル情報の表示を改善しました。また、タグ表示のロジックをutil_ui.dartに移動しました。これにより、コードの可読性と再利用性が向上しました。
